feat: Fix mappers as well

+++ app/module/poll/mapper/VoteMapper.php
@@ -22,7 +22,7 @@ class VoteMapper extends BaseMapper
             Field::int()->withColumn("user_id", true)->setProperty("userId"),
             Field::int()->withColumn("item_id", true)->setProperty("optionId"),
             Field::string(500)->withColumn("text_value")->setProperty("stringValue"),
-            Field::int()->withColumn("numeric_value")->setProperty("numericValue"),
+            Field::float()->withColumn("numeric_value")->setProperty("numericValue"),
             Field::int()->withColumn("boolean_value")->setProperty("booleanValue"),
             Field::int()->withColumn("updated_user_id")->setProperty("updatedById")->setChangeable(false),
             Field::datetime()->withColumn("updated")->setProperty("updatedAt")->setChangeable(false),
